Membuat class baru (Boss.java) dan memindahkan 3 method yang tidak relevan di Employee
